clean up CommandPalette
